Behaviors: Merge screw primitive into arm action (#1371)
* Merge screw primitive into arm action. Cleanup modes.

* Formatting.

* Extract screw primitive code.

* Reorder fields and cleanup.

* Prevent loading crashes.
