+++ ihmc-high-level-behaviors/src/libgdx/java/us/ihmc/rdx/behaviorTree/RDXBehaviorNodeDuplication.java
@@ -13,7 +13,7 @@ import us.ihmc.commons.exception.ExceptionTools;
 import us.ihmc.rdx.behaviorTree.actions.RDXArmAction;
 import us.ihmc.rdx.behaviorTree.actions.RDXNeckAction;
 import us.ihmc.rdx.behaviorTree.actions.RDXSceneAction;
-import us.ihmc.rdx.behaviorTree.actions.RDXScrewPrimitiveAction;
+import us.ihmc.behaviors.behaviorTree.action.actions.ArmActionTaskspaceTrajectoryMode;
 import us.ihmc.rdx.behaviorTree.actions.RDXSpineAction;
 import us.ihmc.rdx.behaviorTree.actions.RDXWalkAction;
 import us.ihmc.rdx.behaviorTree.condition.RDXConditionNode;
@@ -83,13 +83,15 @@ public class RDXBehaviorNodeDuplication
             nominalObjectPose.put("yawInDegrees", -nominalObjectPose.get("yawInDegrees").asDouble());
          }
       }
-      else if (node instanceof RDXArmAction armAction && !armAction.getDefinition().getUsePredefinedJointAngles())
+      else if (node instanceof RDXArmAction armAction && !armAction.getDefinition().getDefinedInJointspace()
+               && armAction.getDefinition().getTaskspaceTrajectoryMode() == ArmActionTaskspaceTrajectoryMode.SINGLE_POSE)
       {
          jsonNode.put("y", -jsonNode.get("y").asDouble()); // TODO: This is specific to the door lever object pose
          jsonNode.put("rollInDegrees", -jsonNode.get("rollInDegrees").asDouble());
          jsonNode.put("yawInDegrees", -jsonNode.get("yawInDegrees").asDouble());
       }
-      else if (node instanceof RDXScrewPrimitiveAction)
+      else if (node instanceof RDXArmAction armAction
+               && armAction.getDefinition().getTaskspaceTrajectoryMode() == ArmActionTaskspaceTrajectoryMode.SCREW_PRIMITIVE)
       {
          if (jsonNode.get("screwAxisPose") instanceof ObjectNode screwAxisPose)
          {
@@ -148,7 +150,7 @@ public class RDXBehaviorNodeDuplication
    private void applyFrameInvariantMirroring(RDXBehaviorTreeNode<?, ?> node, ObjectNode jsonNode)
    {
       if (node instanceof RDXArmAction armAction
-          && armAction.getDefinition().getUsePredefinedJointAngles()
+          && armAction.getDefinition().getDefinedInJointspace()
           && jsonNode.get("preset").asText().equals(ArmActionDefinition.CUSTOM_ANGLES_NAME))
       {
          ArmJointName[] armJointNames = behaviorTree.getRootNode().getSyncedRobot().getRobotModel().getJointMap().getArmJointNames();
