+++ ihmc-high-level-behaviors/src/libgdx/java/us/ihmc/rdx/behaviorTree/actions/RDXArmAction.java
@@ -12,6 +12,7 @@ import us.ihmc.avatar.inverseKinematics.ArmIKSolver;
 import us.ihmc.behaviors.behaviorTree.action.actions.AbilityHandActionState;
 import us.ihmc.behaviors.behaviorTree.action.actions.ArmActionDefinition;
 import us.ihmc.behaviors.behaviorTree.action.actions.ArmActionState;
+import us.ihmc.behaviors.behaviorTree.action.actions.ArmActionTaskspaceTrajectoryMode;
 import us.ihmc.commons.thread.Throttler;
 import us.ihmc.communication.crdt.CRDTBidirectionalRigidBodyTransform;
 import us.ihmc.communication.crdt.CRDTDetachableReferenceFrame;
@@ -32,6 +33,7 @@ import us.ihmc.rdx.imgui.*;
 import us.ihmc.rdx.input.ImGui3DViewInput;
 import us.ihmc.rdx.input.ImGui3DViewPickResult;
 import us.ihmc.rdx.simulation.scs2.RDXRigidBody;
+import us.ihmc.rdx.ui.RDX3DPanel;
 import us.ihmc.rdx.ui.RDX3DPanelTooltip;
 import us.ihmc.rdx.ui.affordances.RDXInteractableHighlightModel;
 import us.ihmc.rdx.ui.affordances.RDXInteractableTools;
@@ -39,6 +41,7 @@ import us.ihmc.rdx.ui.gizmo.RDXSelectablePose3DGizmo;
 import us.ihmc.rdx.ui.graphics.RDXArmMultiBodyGraphic;
 import us.ihmc.rdx.ui.teleoperation.RDXIKSolverColors;
 import us.ihmc.rdx.ui.widgets.ImGuiArmIconWidget;
+import us.ihmc.avatar.drcRobot.ROS2SyncedRobotModel;
 import us.ihmc.robotModels.FullHumanoidRobotModel;
 import us.ihmc.robotics.EuclidCoreMissingTools;
 import us.ihmc.robotics.MultiBodySystemMissingTools;
@@ -59,8 +62,14 @@ public class RDXArmAction extends RDXActionNode<ArmActionState, ArmActionDefinit
 {
    private final ImGuiUniqueLabelMap labels = new ImGuiUniqueLabelMap(getClass());
    private final ImGuiArmIconWidget armIconWidget = new ImGuiArmIconWidget();
+   private final ImBooleanWrapper definedInJointspaceWrapper;
+   private final ImDoubleWrapper trajectoryDurationWidget;
    /** Gizmo is control frame */
    private final RDXSelectablePose3DGizmo poseGizmo;
+   private final ImGuiReferenceFrameLibraryCombo parentFrameComboBox;
+   private final ImBooleanWrapper jointSpaceControlWrapper;
+   private final ImBooleanWrapper holdPoseInWorldLaterWrapper;
+   private final RDXScrewPrimitive screwPrimitive;
    private final SideDependentList<RigidBodyTransformReadOnly> handGraphicToControlFrameTransforms = new SideDependentList<>();
    private final MutableReferenceFrame graphicFrame = new MutableReferenceFrame();
    private final MutableReferenceFrame collisionShapeFrame = new MutableReferenceFrame();
@@ -68,20 +77,16 @@ public class RDXArmAction extends RDXActionNode<ArmActionState, ArmActionDefinit
    private final ImGui3DViewPickResult pickResult = new ImGui3DViewPickResult();
    private final ArrayList<MouseCollidable> mouseCollidables = new ArrayList<>();
    private final SideDependentList<RDXInteractableHighlightModel> highlightModels = new SideDependentList<>();
-   private final ImGuiReferenceFrameLibraryCombo parentFrameComboBox;
-   private final ImDoubleWrapper trajectoryDurationWidget;
    private final String[] configurations = new String[PresetArmConfiguration.values().length + 1];
    private final ImInt currentConfiguration = new ImInt(PresetArmConfiguration.HOME.ordinal() + 1);
    private final ImDoubleWrapper[] jointAngleWidgets = new ImDoubleWrapper[MAX_NUMBER_OF_JOINTS];
    private final SideDependentList<ArmJointName[]> jointNames = new SideDependentList<>();
    private final SideDependentList<double[]> jointLowerLimits = new SideDependentList<>();
    private final SideDependentList<double[]> jointUpperLimits = new SideDependentList<>();
+   private final ImGuiLabelledWidgetAligner widgetAligner = new ImGuiLabelledWidgetAligner();
    private final ImGuiSliderDoubleWrapper linearPositionWeightWidget;
    private final ImGuiSliderDoubleWrapper angularPositionWeightWidget;
    private final ImGuiSliderDoubleWrapper jointspaceWeightWidget;
-   private final ImBooleanWrapper holdPoseInWorldLaterWrapper;
-   private final ImBooleanWrapper jointSpaceControlWrapper;
-   private final ImBooleanWrapper usePredefinedJointAnglesWrapper;
    private final ImDoubleWrapper positionErrorToleranceInput;
    private final ImDoubleWrapper orientationErrorToleranceDegreesInput;
    private final SideDependentList<RDXArmMultiBodyGraphic> armMultiBodyGraphics = new SideDependentList<>();
@@ -97,6 +102,7 @@ public class RDXArmAction extends RDXActionNode<ArmActionState, ArmActionDefinit
 
       poseGizmo = new RDXSelectablePose3DGizmo();
       poseGizmo.create(panel3D);
+      screwPrimitive = new RDXScrewPrimitive(this);
 
       trajectoryDurationWidget = new ImDoubleWrapper(definition::getTrajectoryDuration,
                                                      definition::setTrajectoryDuration,
@@ -104,9 +110,7 @@ public class RDXArmAction extends RDXActionNode<ArmActionState, ArmActionDefinit
       int configurationIndex = 0;
       configurations[configurationIndex++] = ArmActionDefinition.CUSTOM_ANGLES_NAME;
       for (PresetArmConfiguration preset : PresetArmConfiguration.values())
-      {
          configurations[configurationIndex++] = preset.name();
-      }
 
       for (RobotSide side : RobotSide.values)
       {
@@ -155,46 +159,49 @@ public class RDXArmAction extends RDXActionNode<ArmActionState, ArmActionDefinit
       holdPoseInWorldLaterWrapper = new ImBooleanWrapper(definition::getHoldPoseInWorldLater,
                                                          definition::setHoldPoseInWorldLater,
                                                          imBoolean -> ImGui.checkbox(labels.get("Hold pose in world later"), imBoolean));
-      usePredefinedJointAnglesWrapper = new ImBooleanWrapper(definition::getUsePredefinedJointAngles,
-                                                             definition::setUsePredefinedJointAngles,
-                                                             imBoolean ->
-                        {
-                           ImGui.beginDisabled(!state.getPalmFrame().isChildOfWorld());
-                           if (ImGui.checkbox(labels.get("Use Predefined Joint Angles"), imBoolean))
-                           {
-                              definition.setPreset(null); // Preserve joint angles from before
-
-                              if (imBoolean.get()) // When switching into predefined, copy the latest preview joint angles to the definition
-                              {
-                                 for (int i = 0; i < state.getPreviewJointAngles().getLength(); i++)
-                                 {
-                                    definition.getJointAngles().setValue(i, state.getPreviewJointAngles().getValueReadOnly(i));
-                                 }
-                              }
-                              else // When switching from predefined, keep the desired hand in the same place
-                              {
-                                 CRDTDetachableReferenceFrame actionPalmFrame = state.getPalmFrame();
-                                 CRDTBidirectionalRigidBodyTransform palmTransformToParent = definition.getPalmTransformToParent();
-                                 ReferenceFrame previewPalmFrame = armMultiBodyGraphics.get(definition.getSide()).getHandControlFrame();
-                                 FramePose3D previewPalmPose = new FramePose3D();
-                                 previewPalmPose.setToZero(previewPalmFrame);
-                                 previewPalmPose.changeFrame(actionPalmFrame.getReferenceFrame().getParent());
-                                 palmTransformToParent.getValueAndModify().set(previewPalmPose);
-                                 actionPalmFrame.update();
-                              }
-                           }
-                           ImGui.endDisabled();
-                        });
+      definedInJointspaceWrapper = new ImBooleanWrapper(definition::getDefinedInJointspace,
+                                                        definition::setDefinedInJointspace,
+                                                        imBoolean ->
+      {
+         ImGui.text("Definition:");
+         ImGui.sameLine();
+         if (ImGui.radioButton(labels.get("Jointspace"), imBoolean.get()) && !imBoolean.get())
+         {
+            imBoolean.set(true);
+            definition.setPreset(null);
+            for (int i = 0; i < state.getPreviewJointAngles().getLength(); i++)
+               definition.getJointAngles().setValue(i, state.getPreviewJointAngles().getValueReadOnly(i));
+         }
+         ImGui.sameLine();
+         if (ImGui.radioButton(labels.get("Taskspace"), !imBoolean.get()) && imBoolean.get())
+         {
+            imBoolean.set(false);
+            if (definition.getTaskspaceTrajectoryMode() == ArmActionTaskspaceTrajectoryMode.SINGLE_POSE
+                && state.getPalmFrame().isChildOfWorld())
+            {
+               CRDTDetachableReferenceFrame actionPalmFrame = state.getPalmFrame();
+               CRDTBidirectionalRigidBodyTransform palmTransformToParent = definition.getPalmTransformToParent();
+               ReferenceFrame previewPalmFrame = armMultiBodyGraphics.get(definition.getSide()).getHandControlFrame();
+               FramePose3D previewPalmPose = new FramePose3D();
+               previewPalmPose.setToZero(previewPalmFrame);
+               previewPalmPose.changeFrame(actionPalmFrame.getReferenceFrame().getParent());
+               palmTransformToParent.getValueAndModify().set(previewPalmPose);
+               actionPalmFrame.update();
+            }
+         }
+      });
       jointSpaceControlWrapper = new ImBooleanWrapper(definition::getJointspaceOnly,
                                                       definition::setJointspaceOnly,
-                                                      imBoolean -> {
-                                                         if (ImGui.radioButton(labels.get("Hybrid"), !imBoolean.get()))
-                                                            imBoolean.set(false);
-                                                         ImGui.sameLine();
-                                                         if (ImGui.radioButton(labels.get("Jointspace Only"), imBoolean.get()))
-                                                            imBoolean.set(true);
-                                                      });
-      ImGuiLabelledWidgetAligner widgetAligner = new ImGuiLabelledWidgetAligner();
+                                                      imBoolean ->
+      {
+         ImGui.text("Control mode:");
+         ImGui.sameLine();
+         if (ImGui.radioButton(labels.get("Hybrid"), !imBoolean.get()))
+            imBoolean.set(false);
+         ImGui.sameLine();
+         if (ImGui.radioButton(labels.get("Jointspace Only"), imBoolean.get()))
+            imBoolean.set(true);
+      });
       linearPositionWeightWidget = new ImGuiSliderDoubleWrapper("Linear Position Weight", "%.2f", 0.0, 100.0,
                                                                 definition::getLinearPositionWeight,
                                                                 definition::setLinearPositionWeight);
@@ -212,21 +219,21 @@ public class RDXArmAction extends RDXActionNode<ArmActionState, ArmActionDefinit
       jointspaceWeightWidget.addWidgetAligner(widgetAligner);
       positionErrorToleranceInput = new ImDoubleWrapper(() ->
                                                         {
-                                                           if (getDefinition().getUsePredefinedJointAngles())
+                                                           if (definition.getDefinedInJointspace())
                                                               return Math.toDegrees(definition.getPositionErrorTolerance());
                                                            else
                                                               return definition.getPositionErrorTolerance();
                                                         },
                                                         positionErrorTolerance ->
                                                         {
-                                                           if (definition.getUsePredefinedJointAngles())
+                                                           if (definition.getDefinedInJointspace())
                                                               definition.setPositionErrorTolerance(Math.toRadians(positionErrorTolerance));
                                                            else
                                                               definition.setPositionErrorTolerance(positionErrorTolerance);
                                                         },
                                                         imDouble ->
                                                         {
-                                                           if (definition.getUsePredefinedJointAngles())
+                                                           if (definition.getDefinedInJointspace())
                                                               ImGui.inputDouble(labels.get("Position Error Tolerance (%s)".formatted(EuclidCoreMissingTools.DEGREE_SYMBOL)), imDouble);
                                                            else
                                                               ImGui.inputDouble(labels.get("Position Error Tolerance"), imDouble);
@@ -253,9 +260,7 @@ public class RDXArmAction extends RDXActionNode<ArmActionState, ArmActionDefinit
          collisionShapeFrame.update(transformToParent -> transformToParent.set(linkToControlFrameTransform));
 
          for (Collidable handCollidable : handCollidables)
-         {
             mouseCollidables.add(new MouseCollidable(handCollidable));
-         }
 
          armMultiBodyGraphics.put(side, new RDXArmMultiBodyGraphic(robotModel, syncedFullRobotModel, side));
 
@@ -284,32 +289,29 @@ public class RDXArmAction extends RDXActionNode<ArmActionState, ArmActionDefinit
    {
       super.update();
 
-      // IK solution visualization via ghost arms
       showAbilityHand = false;
       if (state.getIsNextForExecution())
          visualizeIK();
 
-      if (definition.getUsePredefinedJointAngles())
+      if (definition.getDefinedInJointspace())
       {
          poseGizmo.setSelected(false);
+         screwPrimitive.deselectGizmos();
 
          PresetArmConfiguration preset = definition.getPreset();
          currentConfiguration.set(preset == null ? 0 : preset.ordinal() + 1);
 
-         // Copy the preset values into the custom data fields so they can be tweaked
-         // relatively when switching to custom angles.
          if (preset != null)
          {
-            // TODO: Would be great if there was a #getPresetArmConfiguration that accepts an array to pack
             double[] jointAngles = syncedRobot.getRobotModel().getPresetArmConfiguration(definition.getSide(), preset);
             for (int i = 0; i < jointAngles.length; i++)
-            {
                definition.getJointAngles().setValue(i, jointAngles[i]);
-            }
          }
       }
-      else if (state.getPalmFrame().isChildOfWorld())
+      else if (isSinglePoseTaskspace() && state.getPalmFrame().isChildOfWorld())
       {
+         screwPrimitive.deselectGizmos();
+
          if (poseGizmo.getPoseGizmo().getGizmoFrame() != state.getPalmFrame().getReferenceFrame())
          {
             poseGizmo.getPoseGizmo().setGizmoFrame(state.getPalmFrame().getReferenceFrame());
@@ -323,30 +325,37 @@ public class RDXArmAction extends RDXActionNode<ArmActionState, ArmActionDefinit
          highlightModels.get(definition.getSide()).setPose(graphicFrame.getReferenceFrame());
 
          if (poseGizmo.isSelected() || isMouseHovering)
-         {
             highlightModels.get(definition.getSide()).setTransparency(0.7);
-         }
          else
-         {
             highlightModels.get(definition.getSide()).setTransparency(0.5);
-         }
+      }
+      else if (isScrewTaskspace())
+      {
+         poseGizmo.setSelected(false);
+         screwPrimitive.update();
       }
    }
 
    private void visualizeIK()
    {
       RDXArmMultiBodyGraphic armMultiBodyGraphic = armMultiBodyGraphics.get(definition.getSide());
-      double solutionQuality = state.getSolutionQuality();
-      if (solutionQuality < ArmIKSolver.GOOD_QUALITY_MAX || lowQualityRenderThrottler.run(0.5))
+
+      if (isScrewTaskspace())
       {
-         armMultiBodyGraphic.getFloatingJoint().getJointPose().set(state.getGoalChestToWorldTransform().getValueReadOnly());
-         for (int i = 0; i < armMultiBodyGraphic.getJoints().length; i++)
-         {
-            armMultiBodyGraphic.getJoints()[i].setQ(state.getPreviewJointAngles().getValueReadOnly(i));
-         }
-         armMultiBodyGraphic.updateAfterModifyingConfiguration();
+         screwPrimitive.visualizeIK(armMultiBodyGraphic);
       }
-      armMultiBodyGraphic.setColor(RDXIKSolverColors.getColor(solutionQuality));
+      else
+      {
+         double solutionQuality = state.getSolutionQuality();
+         if (solutionQuality < ArmIKSolver.GOOD_QUALITY_MAX || lowQualityRenderThrottler.run(0.5))
+         {
+            armMultiBodyGraphic.getFloatingJoint().getJointPose().set(state.getGoalChestToWorldTransform().getValueReadOnly());
+            for (int i = 0; i < armMultiBodyGraphic.getJoints().length; i++)
+               armMultiBodyGraphic.getJoints()[i].setQ(state.getPreviewJointAngles().getValueReadOnly(i));
+            armMultiBodyGraphic.updateAfterModifyingConfiguration();
+         }
+         armMultiBodyGraphic.setColor(RDXIKSolverColors.getColor(solutionQuality));
+      }
 
       if (abilityHands.containsKey(definition.getSide()))
       {
@@ -385,7 +394,7 @@ public class RDXArmAction extends RDXActionNode<ArmActionState, ArmActionDefinit
                else
                   joint = (RevoluteJoint) palm.getChildrenJoints().get(i == 5 ? 4 : i);
                joint.setQ(Math.toRadians(fingerPositions[i]));
-               if (i < 4) // set 2nd joints on non-thumb fingers
+               if (i < 4)
                   ((RevoluteJoint) palm.getChildrenJoints().get(i).getSuccessor().getChildrenJoints().get(0))
                         .setQ(AbilityHandJointName.getQ2Position(Math.toRadians(fingerPositions[i])));
             }
@@ -401,9 +410,14 @@ public class RDXArmAction extends RDXActionNode<ArmActionState, ArmActionDefinit
       super.renderEditableName();
 
       ImGui.sameLine();
-      boolean gizmoWasSelected = poseGizmo.getSelected().get();
-      if (armIconWidget.render(definition.getSide(), gizmoWasSelected, definition.getUsePredefinedJointAngles()))
-         poseGizmo.setSelected(!gizmoWasSelected);
+      if (isScrewTaskspace())
+         ImGui.textDisabled("Screw");
+      else
+      {
+         boolean gizmoWasSelected = poseGizmo.getSelected().get();
+         if (armIconWidget.render(definition.getSide(), gizmoWasSelected, definition.getDefinedInJointspace()))
+            poseGizmo.setSelected(!gizmoWasSelected);
+      }
 
       renderRowEnd();
    }
@@ -411,10 +425,24 @@ public class RDXArmAction extends RDXActionNode<ArmActionState, ArmActionDefinit
    @Override
    protected void renderImGuiWidgetsInternal()
    {
-      trajectoryDurationWidget.renderImGuiWidget();
-      usePredefinedJointAnglesWrapper.renderImGuiWidget();
+      definedInJointspaceWrapper.renderImGuiWidget();
 
-      if (definition.getUsePredefinedJointAngles())
+      if (!definition.getDefinedInJointspace())
+      {
+         ImGui.text("Trajectory mode:");
+         ImGui.sameLine();
+         boolean singlePose = definition.getTaskspaceTrajectoryMode() == ArmActionTaskspaceTrajectoryMode.SINGLE_POSE;
+         if (ImGui.radioButton(labels.get("Single point"), singlePose))
+            definition.setTaskspaceTrajectoryMode(ArmActionTaskspaceTrajectoryMode.SINGLE_POSE);
+         ImGui.sameLine();
+         if (ImGui.radioButton(labels.get("Screw primitive"), !singlePose))
+            definition.setTaskspaceTrajectoryMode(ArmActionTaskspaceTrajectoryMode.SCREW_PRIMITIVE);
+      }
+
+      if (definition.getDefinedInJointspace() || isSinglePoseTaskspace())
+         trajectoryDurationWidget.renderImGuiWidget();
+
+      if (definition.getDefinedInJointspace())
       {
          ImGui.pushItemWidth(200.0f);
          if (ImGui.combo(labels.get("Configuration"), currentConfiguration, configurations))
@@ -426,9 +454,7 @@ public class RDXArmAction extends RDXActionNode<ArmActionState, ArmActionDefinit
             ArmJointName[] armJointNames = syncedRobot.getRobotModel().getJointMap().getArmJointNames(definition.getSide());
             ImGui.pushItemWidth(80.0f);
             for (int i = 0; i < armJointNames.length; i++)
-            {
                jointAngleWidgets[i].renderImGuiWidget();
-            }
             ImGui.popItemWidth();
             if (ImGui.button(labels.get("Set Configuration to Synced Arm")))
             {
@@ -447,8 +473,9 @@ public class RDXArmAction extends RDXActionNode<ArmActionState, ArmActionDefinit
          positionErrorToleranceInput.renderImGuiWidget();
          ImGui.popItemWidth();
       }
-      else
+      else if (isSinglePoseTaskspace())
       {
+         parentFrameComboBox.render();
          ImGui.checkbox(labels.get("Adjust Goal Pose"), poseGizmo.getSelected());
          jointSpaceControlWrapper.renderImGuiWidget();
          if (!definition.getJointspaceOnly())
@@ -456,7 +483,6 @@ public class RDXArmAction extends RDXActionNode<ArmActionState, ArmActionDefinit
             ImGui.sameLine();
             holdPoseInWorldLaterWrapper.renderImGuiWidget();
          }
-         parentFrameComboBox.render();
          if (definition.getJointspaceOnly())
             ImGui.beginDisabled();
          linearPositionWeightWidget.renderImGuiWidget();
@@ -482,89 +508,171 @@ public class RDXArmAction extends RDXActionNode<ArmActionState, ArmActionDefinit
             actionPalmFrame.update();
          }
       }
+      else if (isScrewTaskspace())
+      {
+         screwPrimitive.renderImGuiWidgets();
+      }
    }
 
    @Override
    public void deselectGizmos()
    {
       poseGizmo.setSelected(false);
+      screwPrimitive.deselectGizmos();
    }
 
    public void render3DPanelImGuiOverlays()
    {
       if (isMouseHovering)
-      {
          tooltip.render("%s Action\nIndex: %d\nName: %s".formatted(getLeafTypeTitle(), state.getLeafIndex(), definition.getName()));
-      }
    }
 
    @Override
    public void calculate3DViewPick(ImGui3DViewInput input)
    {
-      if (getSelected() && state.getPalmFrame().isChildOfWorld())
+      if (!getSelected())
+         return;
+
+      if (isSinglePoseTaskspace() && state.getPalmFrame().isChildOfWorld())
       {
          poseGizmo.calculate3DViewPick(input);
 
-         if (!definition.getUsePredefinedJointAngles())
+         pickResult.reset();
+         for (MouseCollidable mouseCollidable : mouseCollidables)
          {
-            pickResult.reset();
-            for (MouseCollidable mouseCollidable : mouseCollidables)
-            {
-               double collision = mouseCollidable.collide(input.getPickRayInWorld(), collisionShapeFrame.getReferenceFrame());
-               if (!Double.isNaN(collision))
-                  pickResult.addPickCollision(collision);
-            }
-            if (pickResult.getPickCollisionWasAddedSinceReset())
-               input.addPickResult(pickResult);
+            double collision = mouseCollidable.collide(input.getPickRayInWorld(), collisionShapeFrame.getReferenceFrame());
+            if (!Double.isNaN(collision))
+               pickResult.addPickCollision(collision);
          }
+         if (pickResult.getPickCollisionWasAddedSinceReset())
+            input.addPickResult(pickResult);
+      }
+      else if (isScrewTaskspace())
+      {
+         screwPrimitive.calculate3DViewPick(input);
       }
    }
 
    @Override
    public void process3DViewInput(ImGui3DViewInput input)
    {
-      if (getSelected() && state.getPalmFrame().isChildOfWorld())
+      if (!getSelected())
+         return;
+
+      if (isSinglePoseTaskspace() && state.getPalmFrame().isChildOfWorld())
       {
          isMouseHovering = input.getClosestPick() == pickResult;
 
          boolean isClickedOn = isMouseHovering && input.mouseReleasedWithoutDrag(ImGuiMouseButton.Left);
          if (isClickedOn)
-         {
             poseGizmo.setSelected(true);
-         }
 
          poseGizmo.process3DViewInput(input, isMouseHovering);
-
          tooltip.setInput(input);
+      }
+      else if (isScrewTaskspace())
+      {
+         screwPrimitive.process3DViewInput(input);
       }
    }
 
    @Override
    public void getRenderables(Array<Renderable> renderables, Pool<Renderable> pool)
    {
-      if ((state.getIsNextForExecution() || getSelected() || state.getIsExecuting()) && state.getPalmFrame().isChildOfWorld())
+      if (!(state.getIsNextForExecution() || getSelected() || state.getIsExecuting()))
+         return;
+
+      if (isSinglePoseTaskspace() && state.getPalmFrame().isChildOfWorld())
       {
-         if (!definition.getUsePredefinedJointAngles() && (getSelected() || poseGizmo.isSelected() || armIconWidget.getIsHovered()))
+         if (getSelected() || poseGizmo.isSelected() || armIconWidget.getIsHovered())
             highlightModels.get(definition.getSide()).getRenderables(renderables, pool);
          poseGizmo.getVirtualRenderables(renderables, pool);
+      }
+      else if (isScrewTaskspace())
+      {
+         screwPrimitive.getRenderables(renderables, pool);
+      }
 
-         if (state.getIsNextForExecution() || state.getIsExecuting())
-         {
-            armMultiBodyGraphics.get(definition.getSide()).getRootBody().getVisualRenderables(renderables, pool);
-            if (showAbilityHand && abilityHands.containsKey(definition.getSide()))
-               abilityHands.get(definition.getSide()).getVisualRenderables(renderables, pool);
-         }
+      if (state.getIsNextForExecution() || state.getIsExecuting())
+      {
+         armMultiBodyGraphics.get(definition.getSide()).getRootBody().getVisualRenderables(renderables, pool);
+         if (showAbilityHand && abilityHands.containsKey(definition.getSide()))
+            abilityHands.get(definition.getSide()).getVisualRenderables(renderables, pool);
       }
    }
 
    @Override
    public String getLeafTypeTitle()
    {
+      if (isScrewTaskspace())
+         return screwPrimitive.getLeafTypeTitle();
       return definition.getSide().getPascalCaseName() + " Arm";
+   }
+
+   private boolean isSinglePoseTaskspace()
+   {
+      return !definition.getDefinedInJointspace()
+             && definition.getTaskspaceTrajectoryMode() == ArmActionTaskspaceTrajectoryMode.SINGLE_POSE;
+   }
+
+   private boolean isScrewTaskspace()
+   {
+      return !definition.getDefinedInJointspace()
+             && definition.getTaskspaceTrajectoryMode() == ArmActionTaskspaceTrajectoryMode.SCREW_PRIMITIVE;
    }
 
    public ReferenceFrame getReferenceFrame()
    {
       return poseGizmo.getPoseGizmo().getGizmoFrame();
+   }
+
+   RDX3DPanel getPanel3D()
+   {
+      return panel3D;
+   }
+
+   ROS2SyncedRobotModel getSyncedRobot()
+   {
+      return syncedRobot;
+   }
+
+   ImGuiUniqueLabelMap getLabels()
+   {
+      return labels;
+   }
+
+   ImGuiLabelledWidgetAligner getWidgetAligner()
+   {
+      return widgetAligner;
+   }
+
+   ImGuiReferenceFrameLibraryCombo getParentFrameComboBox()
+   {
+      return parentFrameComboBox;
+   }
+
+   ImGuiSliderDoubleWrapper getLinearPositionWeightWidget()
+   {
+      return linearPositionWeightWidget;
+   }
+
+   ImGuiSliderDoubleWrapper getAngularPositionWeightWidget()
+   {
+      return angularPositionWeightWidget;
+   }
+
+   ImGuiSliderDoubleWrapper getJointspaceWeightWidget()
+   {
+      return jointspaceWeightWidget;
+   }
+
+   ImDoubleWrapper getPositionErrorToleranceInput()
+   {
+      return positionErrorToleranceInput;
+   }
+
+   ImDoubleWrapper getOrientationErrorToleranceDegreesInput()
+   {
+      return orientationErrorToleranceDegreesInput;
    }
 }
