+++ ihmc-high-level-behaviors/src/main/java/us/ihmc/behaviors/behaviorTree/action/actions/ArmActionExecutor.java
@@ -9,6 +9,8 @@ import ihmc_common_msgs.msg.dds.SE3TrajectoryMessage;
 import ihmc_common_msgs.msg.dds.SE3TrajectoryPointMessage;
 import ihmc_common_msgs.msg.dds.TrajectoryPoint1DMessage;
 import us.ihmc.avatar.inverseKinematics.ArmIKSolver;
+import us.ihmc.avatar.ros2.ROS2ControllerHelper;
+import us.ihmc.avatar.drcRobot.ROS2SyncedRobotModel;
 import us.ihmc.behaviors.behaviorTree.BehaviorTreeRootNodeExecutor;
 import us.ihmc.behaviors.behaviorTree.LeafNodeExecutor;
 import us.ihmc.behaviors.behaviorTree.action.ActionNodeExecutor;
@@ -33,6 +35,7 @@ public class ArmActionExecutor extends ActionNodeExecutor<ArmActionState, ArmAct
    private final FramePose3D goalChestFrame = new FramePose3D();
    private final double[] desiredJointAngles;
    private final ArmJointName[] armJointNames;
+   private final ScrewPrimitiveExecutor screwPrimitive;
 
    public ArmActionExecutor(long id, BehaviorTreeRootNodeExecutor rootNode)
    {
@@ -48,6 +51,8 @@ public class ArmActionExecutor extends ActionNodeExecutor<ArmActionState, ArmAct
       chestAfterJointToPelvis.setToZero(syncedRobot.getReferenceFrames().getChestFrame());
       chestAfterJointToPelvis.changeFrame(syncedRobot.getReferenceFrames().getPelvisFrame());
       chestAfterJointToPelvis.get(chestToPelvisZeroAngles);
+
+      screwPrimitive = new ScrewPrimitiveExecutor(this);
    }
 
    @Override
@@ -57,89 +62,91 @@ public class ArmActionExecutor extends ActionNodeExecutor<ArmActionState, ArmAct
 
       trackingCalculator.update(Conversions.nanosecondsToSeconds(syncedRobot.getTimestamp()));
 
-      if (state.getIsNextForExecution())
-      {
-         SpineActionState concurrentSpineAction = null;
-         PelvisActionState concurrentPelvisAction = null;
-
-         for (int i = state.getExecuteAfterLeafIndex() + 1; i < state.getLeafIndex(); i++)
-         {
-            if (rootNode.getState().getOrderedLeaves().get(i) instanceof SpineActionState spineAction)
-               concurrentSpineAction = spineAction;
-            if (rootNode.getState().getOrderedLeaves().get(i) instanceof PelvisActionState pelvisAction)
-               concurrentPelvisAction = pelvisAction;
-         }
-
-         for (LeafNodeExecutor<?, ?> currentlyExecutingLeaf : rootNode.getCurrentlyExecutingLeaves())
-         {
-            if (currentlyExecutingLeaf.getState() instanceof SpineActionState spineAction)
-               concurrentSpineAction = spineAction;
-            if (currentlyExecutingLeaf.getState() instanceof PelvisActionState pelvisAction)
-               concurrentPelvisAction = pelvisAction;
-         }
-
-         if (concurrentSpineAction == null && concurrentPelvisAction == null)
-         {
-            state.getGoalChestToWorldTransform().accessValue().set(syncedRobot.getReferenceFrames().getChestFrame().getTransformToRoot());
-         }
-         else if (concurrentPelvisAction == null)
-         {
-            concurrentSpineAction.update(); // Ensure state's frames are initialized
-            state.getGoalChestToWorldTransform().accessValue().set(concurrentSpineAction.getChestFrame().getReferenceFrame().getTransformToRoot());
-         }
-         else if (concurrentSpineAction == null)
-         {
-            // FIXME We are ignoring this case for now, just add a pelvis pose to get the desired result
-            //   We need to switch to a proper whole body action node
-            state.getGoalChestToWorldTransform().accessValue().set(syncedRobot.getReferenceFrames().getChestFrame().getTransformToRoot());
-         }
-         else // Combined case
-         {
-            concurrentSpineAction.update(); // Ensure state's frames are initialized
-            concurrentPelvisAction.update(); // Ensure state's frames are initialized
-
-            ReferenceFrame chestActionFrame = concurrentSpineAction.getChestFrame().getReferenceFrame();
-            ReferenceFrame pelvisActionFrame = concurrentPelvisAction.getPelvisFrame().getReferenceFrame();
-
-            chestInPelvis.setToZero(chestActionFrame);
-            chestInPelvis.changeFrame(pelvisActionFrame);
-
-            goalChestFrame.setToZero(pelvisActionFrame);
-            goalChestFrame.getRotation().append(chestInPelvis.getRotation()); // Append chest rotation
-            goalChestFrame.prependTranslation(chestToPelvisZeroAngles.getTranslation());
-            goalChestFrame.changeFrame(ReferenceFrame.getWorldFrame());
-            goalChestFrame.get(state.getGoalChestToWorldTransform().accessValue());
-         }
-         state.getGoalChestFrame().update();
-      }
-
-      if (definition.getUsePredefinedJointAngles())
-      {
-         state.setCanExecute(true);
-
-         state.setSolutionQuality(0.0);
-         for (int i = 0; i < definition.getJointAngles().getLength(); i++)
-         {
-            state.getPreviewJointAngles().accessValue()[i] = definition.getJointAngles().getValueReadOnly(i);
-         }
-      }
+      if (definition.getDefinedInJointspace())
+         updateJointspacePreview();
+      else if (definition.getTaskspaceTrajectoryMode() == ArmActionTaskspaceTrajectoryMode.SCREW_PRIMITIVE)
+         screwPrimitive.updatePreview();
       else
-      {
-         state.setCanExecute(state.getPalmFrame().isChildOfWorld());
-
-         if (state.getPalmFrame().isChildOfWorld() && state.getIsNextForExecution())
-            solveIK();
-      }
+         updateSinglePosePreview();
    }
 
-   private void solveIK()
+   private void updateJointspacePreview()
+   {
+      state.setCanExecute(true);
+      state.setSolutionQuality(0.0);
+      for (int i = 0; i < definition.getJointAngles().getLength(); i++)
+         state.getPreviewJointAngles().accessValue()[i] = definition.getJointAngles().getValueReadOnly(i);
+   }
+
+   private void updateSinglePosePreview()
+   {
+      updateGoalChestTransform();
+      state.setCanExecute(state.getPalmFrame().isChildOfWorld());
+
+      if (state.getPalmFrame().isChildOfWorld() && state.getIsNextForExecution())
+         solveSinglePoseIK();
+   }
+
+   private void updateGoalChestTransform()
+   {
+      if (!state.getIsNextForExecution())
+         return;
+
+      SpineActionState concurrentSpineAction = null;
+      PelvisActionState concurrentPelvisAction = null;
+
+      for (int i = state.getExecuteAfterLeafIndex() + 1; i < state.getLeafIndex(); i++)
+      {
+         if (rootNode.getState().getOrderedLeaves().get(i) instanceof SpineActionState spineAction)
+            concurrentSpineAction = spineAction;
+         if (rootNode.getState().getOrderedLeaves().get(i) instanceof PelvisActionState pelvisAction)
+            concurrentPelvisAction = pelvisAction;
+      }
+
+      for (LeafNodeExecutor<?, ?> currentlyExecutingLeaf : rootNode.getCurrentlyExecutingLeaves())
+      {
+         if (currentlyExecutingLeaf.getState() instanceof SpineActionState spineAction)
+            concurrentSpineAction = spineAction;
+         if (currentlyExecutingLeaf.getState() instanceof PelvisActionState pelvisAction)
+            concurrentPelvisAction = pelvisAction;
+      }
+
+      if (concurrentSpineAction == null && concurrentPelvisAction == null)
+         state.getGoalChestToWorldTransform().accessValue().set(syncedRobot.getReferenceFrames().getChestFrame().getTransformToRoot());
+      else if (concurrentPelvisAction == null)
+      {
+         concurrentSpineAction.update();
+         state.getGoalChestToWorldTransform().accessValue().set(concurrentSpineAction.getChestFrame().getReferenceFrame().getTransformToRoot());
+      }
+      else if (concurrentSpineAction == null)
+         state.getGoalChestToWorldTransform().accessValue().set(syncedRobot.getReferenceFrames().getChestFrame().getTransformToRoot());
+      else
+      {
+         concurrentSpineAction.update();
+         concurrentPelvisAction.update();
+
+         ReferenceFrame chestActionFrame = concurrentSpineAction.getChestFrame().getReferenceFrame();
+         ReferenceFrame pelvisActionFrame = concurrentPelvisAction.getPelvisFrame().getReferenceFrame();
+
+         chestInPelvis.setToZero(chestActionFrame);
+         chestInPelvis.changeFrame(pelvisActionFrame);
+
+         goalChestFrame.setToZero(pelvisActionFrame);
+         goalChestFrame.getRotation().append(chestInPelvis.getRotation());
+         goalChestFrame.prependTranslation(chestToPelvisZeroAngles.getTranslation());
+         goalChestFrame.changeFrame(ReferenceFrame.getWorldFrame());
+         goalChestFrame.get(state.getGoalChestToWorldTransform().accessValue());
+      }
+      state.getGoalChestFrame().update();
+   }
+
+   private void solveSinglePoseIK()
    {
       ArmIKSolver armIKSolver = armIKSolvers.get(definition.getSide());
       armIKSolver.copySourceToWork();
       armIKSolver.update(state.getGoalChestFrame(), state.getPalmFrame().getReferenceFrame());
       armIKSolver.solve();
 
-      // Send the solution back to the UI so the user knows what's gonna happen with the arm.
       state.setSolutionQuality(armIKSolver.getQuality());
       for (int i = 0; i < armIKSolver.getSolutionOneDoFJoints().length; i++)
          state.getPreviewJointAngles().accessValue()[i] = armIKSolver.getSolutionOneDoFJoints()[i].getQ();
@@ -150,63 +157,72 @@ public class ArmActionExecutor extends ActionNodeExecutor<ArmActionState, ArmAct
    {
       super.triggerExecution();
 
+      if (definition.getDefinedInJointspace())
+         triggerJointspaceExecution();
+      else if (definition.getTaskspaceTrajectoryMode() == ArmActionTaskspaceTrajectoryMode.SCREW_PRIMITIVE)
+         screwPrimitive.triggerExecution();
+      else
+         triggerSinglePoseExecution();
+   }
+
+   private void triggerJointspaceExecution()
+   {
+      state.setNominalExecutionDuration(definition.getTrajectoryDuration());
+      trackingCalculator.reset();
+      publishJointspaceCommand(buildJointspaceTrajectoryMessage());
+   }
+
+   private void triggerSinglePoseExecution()
+   {
       state.setNominalExecutionDuration(definition.getTrajectoryDuration());
       trackingCalculator.reset();
 
-      if (definition.getUsePredefinedJointAngles())
-      {
-         JointspaceTrajectoryMessage jointspaceTrajectoryMessage = buildJointspaceTrajectoryMessage();
-         publishJointspaceCommand(jointspaceTrajectoryMessage);
-      }
-      else if (state.getPalmFrame().isChildOfWorld())
-      {
-         solveIK();
-         JointspaceTrajectoryMessage jointspaceTrajectoryMessage = buildJointspaceTrajectoryMessage();
-
-         if (definition.getJointspaceOnly())
-         {
-            publishJointspaceCommand(jointspaceTrajectoryMessage);
-         }
-         else // Publishing taskspace only doesn't work well, so we use hybrid - @dcalvert
-         {
-            ReferenceFrame taskspaceTrajectoryFrame = ReferenceFrame.getWorldFrame();
-            long trajectoryReferenceFrameID = MessageTools.toFrameId(taskspaceTrajectoryFrame);
-            FramePose3D desiredControlFramePose = new FramePose3D(state.getPalmFrame().getReferenceFrame());
-            desiredControlFramePose.changeFrame(taskspaceTrajectoryFrame);
-
-            SE3TrajectoryMessage se3TrajectoryMessage = new SE3TrajectoryMessage();
-            se3TrajectoryMessage.getQueueingProperties().setExecutionMode(QueueableMessage.EXECUTION_MODE_OVERRIDE);
-            se3TrajectoryMessage.getLinearWeightMatrix().setXWeight(definition.getLinearPositionWeight());
-            se3TrajectoryMessage.getLinearWeightMatrix().setYWeight(definition.getLinearPositionWeight());
-            se3TrajectoryMessage.getLinearWeightMatrix().setZWeight(definition.getLinearPositionWeight());
-            se3TrajectoryMessage.getAngularWeightMatrix().setXWeight(definition.getAngularPositionWeight());
-            se3TrajectoryMessage.getAngularWeightMatrix().setYWeight(definition.getAngularPositionWeight());
-            se3TrajectoryMessage.getAngularWeightMatrix().setZWeight(definition.getAngularPositionWeight());
-            se3TrajectoryMessage.getFrameInformation().setTrajectoryReferenceFrameId(trajectoryReferenceFrameID);
-            SE3TrajectoryPointMessage se3TrajectoryPointMessage = se3TrajectoryMessage.getTaskspaceTrajectoryPoints().add();
-            se3TrajectoryPointMessage.setTime(definition.getTrajectoryDuration());
-            se3TrajectoryPointMessage.getPosition().set(desiredControlFramePose.getPosition());
-            se3TrajectoryPointMessage.getOrientation().set(desiredControlFramePose.getOrientation());
-            se3TrajectoryPointMessage.getLinearVelocity().setToZero();
-            se3TrajectoryPointMessage.getAngularVelocity().setToZero();
-
-            HandHybridJointspaceTaskspaceTrajectoryMessage handHybridJointspaceTaskspaceTrajectoryMessage
-                  = new HandHybridJointspaceTaskspaceTrajectoryMessage();
-            handHybridJointspaceTaskspaceTrajectoryMessage.setRobotSide(definition.getSide().toByte());
-            handHybridJointspaceTaskspaceTrajectoryMessage.getTaskspaceTrajectoryMessage().set(se3TrajectoryMessage);
-            handHybridJointspaceTaskspaceTrajectoryMessage.getJointspaceTrajectoryMessage().set(jointspaceTrajectoryMessage);
-            state.getLogger().info("Publishing arm hybrid jointspace taskpace");
-            ros2ControllerHelper.publishToController(handHybridJointspaceTaskspaceTrajectoryMessage);
-         }
-
-         desiredHandControlPose.setFromReferenceFrame(state.getPalmFrame().getReferenceFrame());
-         syncedHandControlPose.setFromReferenceFrame(syncedRobot.getFullRobotModel().getHandControlFrame(definition.getSide()));
-         state.getCommandedTrajectory().setSingleSegmentTrajectory(syncedHandControlPose, desiredHandControlPose, definition.getTrajectoryDuration());
-      }
-      else
+      if (!state.getPalmFrame().isChildOfWorld())
       {
          state.getLogger().error("Cannot execute. Frame is not a child of World frame.");
+         return;
       }
+
+      solveSinglePoseIK();
+      JointspaceTrajectoryMessage jointspaceTrajectoryMessage = buildJointspaceTrajectoryMessage();
+
+      if (definition.getJointspaceOnly())
+         publishJointspaceCommand(jointspaceTrajectoryMessage);
+      else
+      {
+         ReferenceFrame taskspaceTrajectoryFrame = ReferenceFrame.getWorldFrame();
+         long trajectoryReferenceFrameID = MessageTools.toFrameId(taskspaceTrajectoryFrame);
+         FramePose3D desiredControlFramePose = new FramePose3D(state.getPalmFrame().getReferenceFrame());
+         desiredControlFramePose.changeFrame(taskspaceTrajectoryFrame);
+
+         SE3TrajectoryMessage se3TrajectoryMessage = new SE3TrajectoryMessage();
+         se3TrajectoryMessage.getQueueingProperties().setExecutionMode(QueueableMessage.EXECUTION_MODE_OVERRIDE);
+         se3TrajectoryMessage.getLinearWeightMatrix().setXWeight(definition.getLinearPositionWeight());
+         se3TrajectoryMessage.getLinearWeightMatrix().setYWeight(definition.getLinearPositionWeight());
+         se3TrajectoryMessage.getLinearWeightMatrix().setZWeight(definition.getLinearPositionWeight());
+         se3TrajectoryMessage.getAngularWeightMatrix().setXWeight(definition.getAngularPositionWeight());
+         se3TrajectoryMessage.getAngularWeightMatrix().setYWeight(definition.getAngularPositionWeight());
+         se3TrajectoryMessage.getAngularWeightMatrix().setZWeight(definition.getAngularPositionWeight());
+         se3TrajectoryMessage.getFrameInformation().setTrajectoryReferenceFrameId(trajectoryReferenceFrameID);
+         SE3TrajectoryPointMessage se3TrajectoryPointMessage = se3TrajectoryMessage.getTaskspaceTrajectoryPoints().add();
+         se3TrajectoryPointMessage.setTime(definition.getTrajectoryDuration());
+         se3TrajectoryPointMessage.getPosition().set(desiredControlFramePose.getPosition());
+         se3TrajectoryPointMessage.getOrientation().set(desiredControlFramePose.getOrientation());
+         se3TrajectoryPointMessage.getLinearVelocity().setToZero();
+         se3TrajectoryPointMessage.getAngularVelocity().setToZero();
+
+         HandHybridJointspaceTaskspaceTrajectoryMessage handHybridJointspaceTaskspaceTrajectoryMessage
+               = new HandHybridJointspaceTaskspaceTrajectoryMessage();
+         handHybridJointspaceTaskspaceTrajectoryMessage.setRobotSide(definition.getSide().toByte());
+         handHybridJointspaceTaskspaceTrajectoryMessage.getTaskspaceTrajectoryMessage().set(se3TrajectoryMessage);
+         handHybridJointspaceTaskspaceTrajectoryMessage.getJointspaceTrajectoryMessage().set(jointspaceTrajectoryMessage);
+         state.getLogger().info("Publishing arm hybrid jointspace taskpace");
+         ros2ControllerHelper.publishToController(handHybridJointspaceTaskspaceTrajectoryMessage);
+      }
+
+      desiredHandControlPose.setFromReferenceFrame(state.getPalmFrame().getReferenceFrame());
+      syncedHandControlPose.setFromReferenceFrame(syncedRobot.getFullRobotModel().getHandControlFrame(definition.getSide()));
+      state.getCommandedTrajectory().setSingleSegmentTrajectory(syncedHandControlPose, desiredHandControlPose, definition.getTrajectoryDuration());
    }
 
    private void publishJointspaceCommand(JointspaceTrajectoryMessage jointspaceTrajectoryMessage)
@@ -214,7 +230,7 @@ public class ArmActionExecutor extends ActionNodeExecutor<ArmActionState, ArmAct
       ArmTrajectoryMessage armTrajectoryMessage = new ArmTrajectoryMessage();
       armTrajectoryMessage.setRobotSide(definition.getSide().toByte());
       armTrajectoryMessage.getJointspaceTrajectory().set(jointspaceTrajectoryMessage);
-      armTrajectoryMessage.setForceExecution(true); // Prevent the command being rejected because robot is still finishing up walking
+      armTrajectoryMessage.setForceExecution(true);
       state.getLogger().info("Publishing arm jointspace trajectory");
       ros2ControllerHelper.publishToController(armTrajectoryMessage);
    }
@@ -234,67 +250,75 @@ public class ArmActionExecutor extends ActionNodeExecutor<ArmActionState, ArmAct
       {
          state.setIsExecuting(false);
          state.setFailed(true);
-         state.getLogger().error("%s  %s"
-                 .formatted("Position error: %.3f / %.3f".formatted(trackingCalculator.getPositionError(), definition.getPositionErrorTolerance()),
-                            "Orientation error: %.3f / %.3f".formatted(trackingCalculator.getOrientationError(), definition.getOrientationErrorTolerance())));
+         if (!definition.getDefinedInJointspace() && definition.getTaskspaceTrajectoryMode() == ArmActionTaskspaceTrajectoryMode.SINGLE_POSE)
+            state.getLogger().error("%s  %s"
+                    .formatted("Position error: %.3f / %.3f".formatted(trackingCalculator.getPositionError(), definition.getPositionErrorTolerance()),
+                               "Orientation error: %.3f / %.3f".formatted(trackingCalculator.getOrientationError(), definition.getOrientationErrorTolerance())));
          return;
       }
 
-      if (definition.getUsePredefinedJointAngles())
+      if (definition.getDefinedInJointspace())
+         updateJointspaceExecution();
+      else if (definition.getTaskspaceTrajectoryMode() == ArmActionTaskspaceTrajectoryMode.SCREW_PRIMITIVE)
+         screwPrimitive.updateExecution();
+      else
+         updateSinglePoseExecution();
+   }
+
+   private void updateJointspaceExecution()
+   {
+      trackingCalculator.resetJointspaceError();
+
+      for (int i = 0; i < desiredJointAngles.length; i++)
       {
-         trackingCalculator.resetJointspaceError();
+         double desired = desiredJointAngles[i];
+         double current = syncedRobot.getFullRobotModel().getArmJoint(definition.getSide(), armJointNames[i]).getQ();
+         trackingCalculator.addJointData(desired, current);
+      }
+      trackingCalculator.factorInJointspaceErrors(definition.getPositionErrorTolerance());
 
-         for (int i = 0; i < desiredJointAngles.length; i++)
+      if (trackingCalculator.getTimeIsUp())
+      {
+         state.setIsExecuting(false);
+         if (!trackingCalculator.isWithinPositionTolerance())
          {
-            double desired = desiredJointAngles[i];
-            double current = syncedRobot.getFullRobotModel().getArmJoint(definition.getSide(), armJointNames[i]).getQ();
-            trackingCalculator.addJointData(desired, current);
-         }
-         trackingCalculator.factorInJointspaceErrors(definition.getPositionErrorTolerance());
-
-         if (trackingCalculator.getTimeIsUp())
-         {
-            state.setIsExecuting(false);
-            if (!trackingCalculator.isWithinPositionTolerance())
-            {
-               state.getLogger().error("Total jointspace error: %.3f deg".formatted(Math.toDegrees(trackingCalculator.getTotalAbsoluteJointspaceError())));
-               state.setFailed(true);
-            }
+            state.getLogger().error("Total jointspace error: %.3f deg".formatted(Math.toDegrees(trackingCalculator.getTotalAbsoluteJointspaceError())));
+            state.setFailed(true);
          }
       }
-      else if (state.getPalmFrame().isChildOfWorld())
+   }
+
+   private void updateSinglePoseExecution()
+   {
+      if (!state.getPalmFrame().isChildOfWorld())
+         return;
+
+      desiredHandControlPose.setFromReferenceFrame(state.getPalmFrame().getReferenceFrame());
+      syncedHandControlPose.setFromReferenceFrame(syncedRobot.getFullRobotModel().getHandControlFrame(definition.getSide()));
+
+      trackingCalculator.computePoseTrackingData(desiredHandControlPose, syncedHandControlPose);
+      trackingCalculator.factorInR3Errors(definition.getPositionErrorTolerance());
+      trackingCalculator.factoryInSO3Errors(definition.getOrientationErrorTolerance());
+
+      boolean meetsDesiredCompletionCriteria = trackingCalculator.isWithinPositionTolerance();
+      meetsDesiredCompletionCriteria &= trackingCalculator.getTimeIsUp();
+      state.getCurrentPose().accessValue().set(syncedHandControlPose);
+      state.setPositionDistanceToGoalTolerance(definition.getPositionErrorTolerance());
+      state.setOrientationDistanceToGoalTolerance(definition.getOrientationErrorTolerance());
+
+      if (meetsDesiredCompletionCriteria)
       {
-         desiredHandControlPose.setFromReferenceFrame(state.getPalmFrame().getReferenceFrame());
-         syncedHandControlPose.setFromReferenceFrame(syncedRobot.getFullRobotModel().getHandControlFrame(definition.getSide()));
+         state.setIsExecuting(false);
 
-         trackingCalculator.computePoseTrackingData(desiredHandControlPose, syncedHandControlPose);
-         trackingCalculator.factorInR3Errors(definition.getPositionErrorTolerance());
-         trackingCalculator.factoryInSO3Errors(definition.getOrientationErrorTolerance());
-
-         boolean meetsDesiredCompletionCriteria = trackingCalculator.isWithinPositionTolerance();
-         meetsDesiredCompletionCriteria &= trackingCalculator.getTimeIsUp();
-         state.getCurrentPose().accessValue().set(syncedHandControlPose);
-         state.setPositionDistanceToGoalTolerance(definition.getPositionErrorTolerance());
-         state.setOrientationDistanceToGoalTolerance(definition.getOrientationErrorTolerance());
-
-         if (meetsDesiredCompletionCriteria)
-         {
-            state.setIsExecuting(false);
-
-            if (!definition.getJointspaceOnly() && !definition.getHoldPoseInWorldLater())
-            {
-               disengageHoldPoseInWorld();
-            }
-         }
+         if (!definition.getJointspaceOnly() && !definition.getHoldPoseInWorldLater())
+            disengageHoldPoseInWorld();
       }
    }
 
    private void disengageHoldPoseInWorld()
    {
-      JointspaceTrajectoryMessage jointspaceTrajectoryMessage = buildJointspaceTrajectoryMessage();
-
       state.getLogger().info("Disengaging holding hand in taskspace");
-      publishJointspaceCommand(jointspaceTrajectoryMessage);
+      publishJointspaceCommand(buildJointspaceTrajectoryMessage());
    }
 
    private JointspaceTrajectoryMessage buildJointspaceTrajectoryMessage()
@@ -302,7 +326,7 @@ public class ArmActionExecutor extends ActionNodeExecutor<ArmActionState, ArmAct
       JointspaceTrajectoryMessage jointspaceTrajectoryMessage = new JointspaceTrajectoryMessage();
       jointspaceTrajectoryMessage.getQueueingProperties().setExecutionMode(QueueableMessage.EXECUTION_MODE_OVERRIDE);
 
-      if (definition.getUsePredefinedJointAngles())
+      if (definition.getDefinedInJointspace())
          for (int i = 0; i < desiredJointAngles.length; i++)
             desiredJointAngles[i] = definition.getJointAngles().getValueReadOnly(i);
       else
@@ -321,5 +345,25 @@ public class ArmActionExecutor extends ActionNodeExecutor<ArmActionState, ArmAct
       }
 
       return jointspaceTrajectoryMessage;
+   }
+
+   SideDependentList<ArmIKSolver> getArmIKSolvers()
+   {
+      return armIKSolvers;
+   }
+
+   TrajectoryTrackingErrorCalculator getTrackingCalculator()
+   {
+      return trackingCalculator;
+   }
+
+   ROS2SyncedRobotModel getSyncedRobot()
+   {
+      return syncedRobot;
+   }
+
+   ROS2ControllerHelper getRos2ControllerHelper()
+   {
+      return ros2ControllerHelper;
    }
 }
