+++ ihmc-high-level-behaviors/src/main/java/us/ihmc/behaviors/behaviorTree/action/actions/ArmActionState.java
@@ -6,6 +6,7 @@ import us.ihmc.behaviors.behaviorTree.action.ActionNodeState;
 import us.ihmc.communication.crdt.*;
 import us.ihmc.communication.ros2.ROS2ActorDesignation;
 import us.ihmc.euclid.referenceFrame.ReferenceFrame;
+import us.ihmc.robotics.referenceFrames.DetachableReferenceFrame;
 import us.ihmc.robotics.referenceFrames.ReferenceFrameMissingTools;
 import us.ihmc.robotics.robotSide.RobotSide;
 import us.ihmc.robotics.robotSide.SideDependentList;
@@ -22,10 +23,11 @@ public class ArmActionState extends ActionNodeState<ArmActionDefinition>
     */
    private final CRDTStatusRigidBodyTransform goalChestToWorldTransform;
    private final ReferenceFrame goalChestFrame;
-   private final CRDTStatusVector3D force;
-   private final CRDTStatusVector3D torque;
    private final CRDTStatusDoubleArray previewJointAngles;
    private final CRDTStatusDouble solutionQuality;
+   private final ScrewPrimitiveState screwPrimitive;
+   private final CRDTStatusVector3D force;
+   private final CRDTStatusVector3D torque;
    private final SideDependentList<Integer> numberOfJoints = new SideDependentList<>();
 
    public ArmActionState(long id, BehaviorTreeRootNodeState rootNode)
@@ -38,10 +40,11 @@ public class ArmActionState extends ActionNodeState<ArmActionDefinition>
       goalChestToWorldTransform = new CRDTStatusRigidBodyTransform(ROS2ActorDesignation.ROBOT, crdtInfo);
       goalChestFrame = ReferenceFrameMissingTools.constructFrameWithChangingTransformToParent(ReferenceFrame.getWorldFrame(),
                                                                                               goalChestToWorldTransform.getValueReadOnly());
-      force = new CRDTStatusVector3D(ROS2ActorDesignation.ROBOT, crdtInfo);
-      torque = new CRDTStatusVector3D(ROS2ActorDesignation.ROBOT, crdtInfo);
       previewJointAngles = new CRDTStatusDoubleArray(ROS2ActorDesignation.ROBOT, crdtInfo, MAX_NUMBER_OF_JOINTS);
       solutionQuality = new CRDTStatusDouble(ROS2ActorDesignation.ROBOT, crdtInfo, Double.NaN);
+      screwPrimitive = new ScrewPrimitiveState(scene, crdtInfo, definition.getScrewPrimitive());
+      force = new CRDTStatusVector3D(ROS2ActorDesignation.ROBOT, crdtInfo);
+      torque = new CRDTStatusVector3D(ROS2ActorDesignation.ROBOT, crdtInfo);
 
       for (RobotSide side : RobotSide.values)
          numberOfJoints.put(side, robotModel.getJointMap().getArmJointNamesAsStrings(side).size());
@@ -51,6 +54,7 @@ public class ArmActionState extends ActionNodeState<ArmActionDefinition>
    public void update()
    {
       palmFrame.update();
+      screwPrimitive.update(definition.getPalmParentFrameName());
    }
 
    @Override
@@ -58,10 +62,11 @@ public class ArmActionState extends ActionNodeState<ArmActionDefinition>
    {
       boolean hasStatus = super.hasStatus();
       hasStatus |= goalChestToWorldTransform.pollHasStatus();
-      hasStatus |= force.pollHasStatus();
-      hasStatus |= torque.pollHasStatus();
       hasStatus |= previewJointAngles.pollHasStatus();
       hasStatus |= solutionQuality.pollHasStatus();
+      hasStatus |= screwPrimitive.pollHasStatus();
+      hasStatus |= force.pollHasStatus();
+      hasStatus |= torque.pollHasStatus();
       return hasStatus;
    }
 
@@ -72,13 +77,11 @@ public class ArmActionState extends ActionNodeState<ArmActionDefinition>
       super.toMessage(message.getState());
 
       goalChestToWorldTransform.toMessage(message.getGoalChestTransformToWorld());
+      previewJointAngles.toMessage(message.getJointAngles());
+      message.setSolutionQuality(solutionQuality.toMessage());
+      screwPrimitive.toMessage(message);
       force.toMessage(message.getForce());
       torque.toMessage(message.getTorque());
-      for (int i = 0; i < MAX_NUMBER_OF_JOINTS; i++)
-      {
-         previewJointAngles.toMessage(message.getJointAngles());
-      }
-      message.setSolutionQuality(solutionQuality.toMessage());
    }
 
    public void fromMessage(ArmActionStateMessage message)
@@ -87,17 +90,28 @@ public class ArmActionState extends ActionNodeState<ArmActionDefinition>
 
       super.fromMessage(message.getState());
 
-      force.fromMessage(message.getForce());
-      torque.fromMessage(message.getTorque());
-      previewJointAngles.fromMessage(message.getJointAngles());
-      solutionQuality.fromMessage(message.getSolutionQuality());
       goalChestToWorldTransform.fromMessage(message.getGoalChestTransformToWorld());
       goalChestFrame.update();
+      previewJointAngles.fromMessage(message.getJointAngles());
+      solutionQuality.fromMessage(message.getSolutionQuality());
+      screwPrimitive.fromMessage(message);
+      force.fromMessage(message.getForce());
+      torque.fromMessage(message.getTorque());
    }
 
    public CRDTDetachableReferenceFrame getPalmFrame()
    {
       return palmFrame;
+   }
+
+   public ScrewPrimitiveState getScrewPrimitive()
+   {
+      return screwPrimitive;
+   }
+
+   public DetachableReferenceFrame getScrewFrame()
+   {
+      return screwPrimitive.getScrewFrame();
    }
 
    public CRDTStatusRigidBodyTransform getGoalChestToWorldTransform()
@@ -123,6 +137,41 @@ public class ArmActionState extends ActionNodeState<ArmActionDefinition>
    public CRDTStatusDoubleArray getPreviewJointAngles()
    {
       return previewJointAngles;
+   }
+
+   public CRDTStatusPoseList getPreviewTrajectory()
+   {
+      return screwPrimitive.getPreviewTrajectory();
+   }
+
+   public CRDTStatusDouble getPreviewTrajectoryDuration()
+   {
+      return screwPrimitive.getPreviewTrajectoryDuration();
+   }
+
+   public CRDTStatusDouble getPreviewTrajectoryLinearVelocity()
+   {
+      return screwPrimitive.getPreviewTrajectoryLinearVelocity();
+   }
+
+   public CRDTStatusDouble getPreviewTrajectoryAngularVelocity()
+   {
+      return screwPrimitive.getPreviewTrajectoryAngularVelocity();
+   }
+
+   public CRDTStatusDouble getPreviewRequestedTime()
+   {
+      return screwPrimitive.getPreviewRequestedTime();
+   }
+
+   public CRDTStatusDoubleArray getScrewPreviewJointAngles()
+   {
+      return screwPrimitive.getScrewPreviewJointAngles();
+   }
+
+   public CRDTStatusDouble getPreviewSolutionQuality()
+   {
+      return screwPrimitive.getPreviewSolutionQuality();
    }
 
    public int getNumberOfJoints()
