+++ ihmc-high-level-behaviors/src/main/java/us/ihmc/behaviors/behaviorTree/action/actions/ArmActionTaskspaceTrajectoryMode.java
@@ -1,0 +1,9 @@
+package us.ihmc.behaviors.behaviorTree.action.actions;
+
+public enum ArmActionTaskspaceTrajectoryMode
+{
+   SINGLE_POSE,
+   SCREW_PRIMITIVE;
+
+   public static final ArmActionTaskspaceTrajectoryMode[] values = values();
+}
