+++ ihmc-interfaces/src/main/generated-idl/behavior_msgs/msg/ArmActionDefinitionMessage_.idl
@@ -23,21 +23,17 @@ module behavior_msgs
         @defaultValue(value=255)
         octet robot_side;
         /**
-         * Name of the frame the this action is expressed in
+         * Whether the action is defined in jointspace (true) or taskspace (false)
          */
-        string parent_frame_name;
+        boolean defined_in_jointspace;
         /**
-         * Transform that expresses the hand pose in the parent frame
+         * Taskspace trajectory mode: 0 = single pose, 1 = screw primitive
          */
-        controller_msgs::msg::dds::RigidBodyTransformMessage transform_to_parent;
+        octet taskspace_trajectory_mode;
         /**
-         * Whether the rigid body is controlled in jointspace (true) or taskspace (false)
+         * The trajectory duration
          */
-        boolean joint_space_control;
-        /**
-         * Whether to use predefined joint angles
-         */
-        boolean use_predefined_joint_angles;
+        double trajectory_duration;
         /**
          * Preset arm configuration
          */
@@ -47,13 +43,29 @@ module behavior_msgs
          */
         double joint_angles[7];
         /**
-         * The trajectory duration
+         * Name of the frame the this action is expressed in
          */
-        double trajectory_duration;
+        string parent_frame_name;
+        /**
+         * Transform that expresses the hand pose in the parent frame
+         */
+        controller_msgs::msg::dds::RigidBodyTransformMessage transform_to_parent;
+        /**
+         * Whether the rigid body is controlled in jointspace (true) or hybrid (false) during single-point taskspace execution
+         */
+        boolean joint_space_control;
         /**
          * Whether maintaining the rigid body controlled in world after the action is complete
          */
         boolean hold_pose_in_world;
+        /**
+         * Screw primitive fields
+         */
+        controller_msgs::msg::dds::RigidBodyTransformMessage screw_axis_pose;
+        double translation;
+        double rotation;
+        double max_linear_velocity;
+        double max_angular_velocity;
         double linear_position_weight;
         double angular_position_weight;
         double jointspace_weight;
