+++ ihmc-interfaces/src/main/generated-idl/behavior_msgs/msg/ArmActionStateMessage_.idl
@@ -4,6 +4,7 @@
 #include "behavior_msgs/msg/./ActionNodeStateMessage_.idl"
 #include "behavior_msgs/msg/./ArmActionDefinitionMessage_.idl"
 #include "controller_msgs/msg/./RigidBodyTransformMessage_.idl"
+#include "geometry_msgs/msg/./Pose_.idl"
 #include "geometry_msgs/msg/./Vector3_.idl"
 module behavior_msgs
 {
@@ -28,14 +29,6 @@ module behavior_msgs
          */
         controller_msgs::msg::dds::RigidBodyTransformMessage goal_chest_transform_to_world;
         /**
-         * Hand force
-         */
-        geometry_msgs::msg::dds::Vector3 force;
-        /**
-         * Hand torque
-         */
-        geometry_msgs::msg::dds::Vector3 torque;
-        /**
          * Joint angles
          */
         double joint_angles[7];
@@ -43,6 +36,27 @@ module behavior_msgs
          * Quality of the IK solution
          */
         double solution_quality;
+        /**
+         * Screw primitive preview fields
+         */
+        sequence<geometry_msgs::msg::dds::Pose, 50> preview_trajectory;
+        double preview_trajectory_duration;
+        double preview_trajectory_linear_velocity;
+        double preview_trajectory_angular_velocity;
+        /**
+         * The user requested IK preview time normalized from 0.0 to 1.0
+         */
+        double preview_requested_time;
+        double preview_joint_angles[7];
+        double preview_solution_quality;
+        /**
+         * Hand force
+         */
+        geometry_msgs::msg::dds::Vector3 force;
+        /**
+         * Hand torque
+         */
+        geometry_msgs::msg::dds::Vector3 torque;
       };
     };
   };
