--- ihmc-interfaces/src/main/generated-idl/behavior_msgs/msg/ScrewPrimitiveActionDefinitionMessage_.idl
@@ -1,63 +1,0 @@
-#ifndef __behavior_msgs__msg__ScrewPrimitiveActionDefinitionMessage__idl__
-#define __behavior_msgs__msg__ScrewPrimitiveActionDefinitionMessage__idl__
-
-#include "behavior_msgs/msg/./ActionNodeDefinitionMessage_.idl"
-#include "controller_msgs/msg/./RigidBodyTransformMessage_.idl"
-module behavior_msgs
-{
-  module msg
-  {
-    module dds
-    {
-
-      @TypeCode(type="behavior_msgs::msg::dds_::ScrewPrimitiveActionDefinitionMessage_")
-      struct ScrewPrimitiveActionDefinitionMessage
-      {
-        /**
-         * Parent definition fields
-         */
-        behavior_msgs::msg::dds::ActionNodeDefinitionMessage definition;
-        /**
-         * Specifies the side of the robot that this message refers to.
-         */
-        @defaultValue(value=255)
-        octet robot_side;
-        /**
-         * Name of the frame the this action is expressed in
-         */
-        string object_frame_name;
-        /**
-         * The pose of the screw axis in the object frame
-         */
-        controller_msgs::msg::dds::RigidBodyTransformMessage screw_axis_pose;
-        /**
-         * The magnitude of the translation component
-         */
-        double translation;
-        /**
-         * The magnitude of the rotation component
-         */
-        double rotation;
-        /**
-         * The max linear velocity
-         */
-        double max_linear_velocity;
-        /**
-         * The max angular velocity
-         */
-        double max_angular_velocity;
-        /**
-         * Whether the trajectory is controlled in jointspace (true) or hybrid jointspace and taskspace (false)
-         */
-        boolean jointspace_only;
-        double linear_position_weight;
-        double angular_position_weight;
-        double jointspace_weight;
-        double position_error_tolerance;
-        double orientation_error_tolerance;
-      };
-    };
-  };
-};
-
-#endif
