--- ihmc-interfaces/src/main/generated-idl/behavior_msgs/msg/ScrewPrimitiveActionStateMessage_.idl
@@ -1,49 +1,0 @@
-#ifndef __behavior_msgs__msg__ScrewPrimitiveActionStateMessage__idl__
-#define __behavior_msgs__msg__ScrewPrimitiveActionStateMessage__idl__
-
-#include "behavior_msgs/msg/./ActionNodeStateMessage_.idl"
-#include "behavior_msgs/msg/./ScrewPrimitiveActionDefinitionMessage_.idl"
-#include "geometry_msgs/msg/./Pose_.idl"
-#include "geometry_msgs/msg/./Vector3_.idl"
-module behavior_msgs
-{
-  module msg
-  {
-    module dds
-    {
-
-      @TypeCode(type="behavior_msgs::msg::dds_::ScrewPrimitiveActionStateMessage_")
-      struct ScrewPrimitiveActionStateMessage
-      {
-        /**
-         * Parent state fields
-         */
-        behavior_msgs::msg::dds::ActionNodeStateMessage state;
-        /**
-         * Definition
-         */
-        behavior_msgs::msg::dds::ScrewPrimitiveActionDefinitionMessage definition;
-        sequence<geometry_msgs::msg::dds::Pose, 50> preview_trajectory;
-        /**
-         * Hand force
-         */
-        geometry_msgs::msg::dds::Vector3 force;
-        /**
-         * Hand torque
-         */
-        geometry_msgs::msg::dds::Vector3 torque;
-        double preview_trajectory_duration;
-        double preview_trajectory_linear_velocity;
-        double preview_trajectory_angular_velocity;
-        /**
-         * The user requested IK preview time normalized from 0.0 to 1.0
-         */
-        double preview_requested_time;
-        double preview_joint_angles[7];
-        double preview_solution_quality;
-      };
-    };
-  };
-};
-
-#endif
