+++ ihmc-interfaces/src/main/generated-java/behavior_msgs/msg/dds/ArmActionDefinitionMessagePubSubType.java
@@ -15,7 +15,7 @@ public class ArmActionDefinitionMessagePubSubType implements us.ihmc.pubsub.Topi
    @Override
    public final java.lang.String getDefinitionChecksum()
    {
-   		return "330f2ec5c5738c622ba345b448f9569e40a8e8c75ae99bb88a861fc083b16307";
+   		return "fbcd056ccbdd0293bfbf8f37b9de8b085efddb336821c74173925e52f0d758b3";
    }
    
    @Override
@@ -56,6 +56,16 @@ public class ArmActionDefinitionMessagePubSubType implements us.ihmc.pubsub.Topi
 
       current_alignment += 1 + us.ihmc.idl.CDR.alignment(current_alignment, 1);
 
+      current_alignment += 1 + us.ihmc.idl.CDR.alignment(current_alignment, 1);
+
+      current_alignment += 1 + us.ihmc.idl.CDR.alignment(current_alignment, 1);
+
+      current_alignment += 8 + us.ihmc.idl.CDR.alignment(current_alignment, 8);
+
+      current_alignment += 1 + us.ihmc.idl.CDR.alignment(current_alignment, 1);
+
+      current_alignment += ((7) * 8) + us.ihmc.idl.CDR.alignment(current_alignment, 8);
+
       current_alignment += 4 + us.ihmc.idl.CDR.alignment(current_alignment, 4) + 255 + 1;
       current_alignment += controller_msgs.msg.dds.RigidBodyTransformMessagePubSubType.getMaxCdrSerializedSize(current_alignment);
 
@@ -63,13 +73,15 @@ public class ArmActionDefinitionMessagePubSubType implements us.ihmc.pubsub.Topi
 
       current_alignment += 1 + us.ihmc.idl.CDR.alignment(current_alignment, 1);
 
-      current_alignment += 1 + us.ihmc.idl.CDR.alignment(current_alignment, 1);
-
-      current_alignment += ((7) * 8) + us.ihmc.idl.CDR.alignment(current_alignment, 8);
+      current_alignment += controller_msgs.msg.dds.RigidBodyTransformMessagePubSubType.getMaxCdrSerializedSize(current_alignment);
 
       current_alignment += 8 + us.ihmc.idl.CDR.alignment(current_alignment, 8);
 
-      current_alignment += 1 + us.ihmc.idl.CDR.alignment(current_alignment, 1);
+      current_alignment += 8 + us.ihmc.idl.CDR.alignment(current_alignment, 8);
+
+      current_alignment += 8 + us.ihmc.idl.CDR.alignment(current_alignment, 8);
+
+      current_alignment += 8 + us.ihmc.idl.CDR.alignment(current_alignment, 8);
 
       current_alignment += 8 + us.ihmc.idl.CDR.alignment(current_alignment, 8);
 
@@ -99,6 +111,19 @@ public class ArmActionDefinitionMessagePubSubType implements us.ihmc.pubsub.Topi
       current_alignment += 1 + us.ihmc.idl.CDR.alignment(current_alignment, 1);
 
 
+      current_alignment += 1 + us.ihmc.idl.CDR.alignment(current_alignment, 1);
+
+
+      current_alignment += 1 + us.ihmc.idl.CDR.alignment(current_alignment, 1);
+
+
+      current_alignment += 8 + us.ihmc.idl.CDR.alignment(current_alignment, 8);
+
+
+      current_alignment += 1 + us.ihmc.idl.CDR.alignment(current_alignment, 1);
+
+
+      current_alignment += ((7) * 8) + us.ihmc.idl.CDR.alignment(current_alignment, 8);
       current_alignment += 4 + us.ihmc.idl.CDR.alignment(current_alignment, 4) + data.getParentFrameName().length() + 1;
 
       current_alignment += controller_msgs.msg.dds.RigidBodyTransformMessagePubSubType.getCdrSerializedSize(data.getTransformToParent(), current_alignment);
@@ -109,14 +134,18 @@ public class ArmActionDefinitionMessagePubSubType implements us.ihmc.pubsub.Topi
       current_alignment += 1 + us.ihmc.idl.CDR.alignment(current_alignment, 1);
 
 
-      current_alignment += 1 + us.ihmc.idl.CDR.alignment(current_alignment, 1);
+      current_alignment += controller_msgs.msg.dds.RigidBodyTransformMessagePubSubType.getCdrSerializedSize(data.getScrewAxisPose(), current_alignment);
 
-
-      current_alignment += ((7) * 8) + us.ihmc.idl.CDR.alignment(current_alignment, 8);
       current_alignment += 8 + us.ihmc.idl.CDR.alignment(current_alignment, 8);
 
 
-      current_alignment += 1 + us.ihmc.idl.CDR.alignment(current_alignment, 1);
+      current_alignment += 8 + us.ihmc.idl.CDR.alignment(current_alignment, 8);
+
+
+      current_alignment += 8 + us.ihmc.idl.CDR.alignment(current_alignment, 8);
+
+
+      current_alignment += 8 + us.ihmc.idl.CDR.alignment(current_alignment, 8);
 
 
       current_alignment += 8 + us.ihmc.idl.CDR.alignment(current_alignment, 8);
@@ -143,14 +172,11 @@ public class ArmActionDefinitionMessagePubSubType implements us.ihmc.pubsub.Topi
       behavior_msgs.msg.dds.ActionNodeDefinitionMessagePubSubType.write(data.getDefinition(), cdr);
       cdr.write_type_9(data.getRobotSide());
 
-      if(data.getParentFrameName().length() <= 255)
-      cdr.write_type_d(data.getParentFrameName());else
-          throw new RuntimeException("parent_frame_name field exceeds the maximum length: %d > %d".formatted(data.getParentFrameName().length(), 255));
+      cdr.write_type_7(data.getDefinedInJointspace());
 
-      controller_msgs.msg.dds.RigidBodyTransformMessagePubSubType.write(data.getTransformToParent(), cdr);
-      cdr.write_type_7(data.getJointSpaceControl());
+      cdr.write_type_9(data.getTaskspaceTrajectoryMode());
 
-      cdr.write_type_7(data.getUsePredefinedJointAngles());
+      cdr.write_type_6(data.getTrajectoryDuration());
 
       cdr.write_type_9(data.getPreset());
 
@@ -159,9 +185,23 @@ public class ArmActionDefinitionMessagePubSubType implements us.ihmc.pubsub.Topi
         	cdr.write_type_6(data.getJointAngles()[i0]);	
       }
 
-      cdr.write_type_6(data.getTrajectoryDuration());
+      if(data.getParentFrameName().length() <= 255)
+      cdr.write_type_d(data.getParentFrameName());else
+          throw new RuntimeException("parent_frame_name field exceeds the maximum length: %d > %d".formatted(data.getParentFrameName().length(), 255));
+
+      controller_msgs.msg.dds.RigidBodyTransformMessagePubSubType.write(data.getTransformToParent(), cdr);
+      cdr.write_type_7(data.getJointSpaceControl());
 
       cdr.write_type_7(data.getHoldPoseInWorld());
+
+      controller_msgs.msg.dds.RigidBodyTransformMessagePubSubType.write(data.getScrewAxisPose(), cdr);
+      cdr.write_type_6(data.getTranslation());
+
+      cdr.write_type_6(data.getRotation());
+
+      cdr.write_type_6(data.getMaxLinearVelocity());
+
+      cdr.write_type_6(data.getMaxAngularVelocity());
 
       cdr.write_type_6(data.getLinearPositionWeight());
 
@@ -180,11 +220,11 @@ public class ArmActionDefinitionMessagePubSubType implements us.ihmc.pubsub.Topi
       behavior_msgs.msg.dds.ActionNodeDefinitionMessagePubSubType.read(data.getDefinition(), cdr);	
       data.setRobotSide(cdr.read_type_9());
       	
-      cdr.read_type_d(data.getParentFrameName());	
-      controller_msgs.msg.dds.RigidBodyTransformMessagePubSubType.read(data.getTransformToParent(), cdr);	
-      data.setJointSpaceControl(cdr.read_type_7());
+      data.setDefinedInJointspace(cdr.read_type_7());
       	
-      data.setUsePredefinedJointAngles(cdr.read_type_7());
+      data.setTaskspaceTrajectoryMode(cdr.read_type_9());
+      	
+      data.setTrajectoryDuration(cdr.read_type_6());
       	
       data.setPreset(cdr.read_type_9());
       	
@@ -194,9 +234,20 @@ public class ArmActionDefinitionMessagePubSubType implements us.ihmc.pubsub.Topi
         	
       }
       	
-      data.setTrajectoryDuration(cdr.read_type_6());
+      cdr.read_type_d(data.getParentFrameName());	
+      controller_msgs.msg.dds.RigidBodyTransformMessagePubSubType.read(data.getTransformToParent(), cdr);	
+      data.setJointSpaceControl(cdr.read_type_7());
       	
       data.setHoldPoseInWorld(cdr.read_type_7());
+      	
+      controller_msgs.msg.dds.RigidBodyTransformMessagePubSubType.read(data.getScrewAxisPose(), cdr);	
+      data.setTranslation(cdr.read_type_6());
+      	
+      data.setRotation(cdr.read_type_6());
+      	
+      data.setMaxLinearVelocity(cdr.read_type_6());
+      	
+      data.setMaxAngularVelocity(cdr.read_type_6());
       	
       data.setLinearPositionWeight(cdr.read_type_6());
       	
@@ -217,15 +268,22 @@ public class ArmActionDefinitionMessagePubSubType implements us.ihmc.pubsub.Topi
       ser.write_type_a("definition", new behavior_msgs.msg.dds.ActionNodeDefinitionMessagePubSubType(), data.getDefinition());
 
       ser.write_type_9("robot_side", data.getRobotSide());
+      ser.write_type_7("defined_in_jointspace", data.getDefinedInJointspace());
+      ser.write_type_9("taskspace_trajectory_mode", data.getTaskspaceTrajectoryMode());
+      ser.write_type_6("trajectory_duration", data.getTrajectoryDuration());
+      ser.write_type_9("preset", data.getPreset());
+      ser.write_type_f("joint_angles", data.getJointAngles());
       ser.write_type_d("parent_frame_name", data.getParentFrameName());
       ser.write_type_a("transform_to_parent", new controller_msgs.msg.dds.RigidBodyTransformMessagePubSubType(), data.getTransformToParent());
 
       ser.write_type_7("joint_space_control", data.getJointSpaceControl());
-      ser.write_type_7("use_predefined_joint_angles", data.getUsePredefinedJointAngles());
-      ser.write_type_9("preset", data.getPreset());
-      ser.write_type_f("joint_angles", data.getJointAngles());
-      ser.write_type_6("trajectory_duration", data.getTrajectoryDuration());
       ser.write_type_7("hold_pose_in_world", data.getHoldPoseInWorld());
+      ser.write_type_a("screw_axis_pose", new controller_msgs.msg.dds.RigidBodyTransformMessagePubSubType(), data.getScrewAxisPose());
+
+      ser.write_type_6("translation", data.getTranslation());
+      ser.write_type_6("rotation", data.getRotation());
+      ser.write_type_6("max_linear_velocity", data.getMaxLinearVelocity());
+      ser.write_type_6("max_angular_velocity", data.getMaxAngularVelocity());
       ser.write_type_6("linear_position_weight", data.getLinearPositionWeight());
       ser.write_type_6("angular_position_weight", data.getAngularPositionWeight());
       ser.write_type_6("jointspace_weight", data.getJointspaceWeight());
@@ -239,15 +297,22 @@ public class ArmActionDefinitionMessagePubSubType implements us.ihmc.pubsub.Topi
       ser.read_type_a("definition", new behavior_msgs.msg.dds.ActionNodeDefinitionMessagePubSubType(), data.getDefinition());
 
       data.setRobotSide(ser.read_type_9("robot_side"));
+      data.setDefinedInJointspace(ser.read_type_7("defined_in_jointspace"));
+      data.setTaskspaceTrajectoryMode(ser.read_type_9("taskspace_trajectory_mode"));
+      data.setTrajectoryDuration(ser.read_type_6("trajectory_duration"));
+      data.setPreset(ser.read_type_9("preset"));
+      ser.read_type_f("joint_angles", data.getJointAngles());
       ser.read_type_d("parent_frame_name", data.getParentFrameName());
       ser.read_type_a("transform_to_parent", new controller_msgs.msg.dds.RigidBodyTransformMessagePubSubType(), data.getTransformToParent());
 
       data.setJointSpaceControl(ser.read_type_7("joint_space_control"));
-      data.setUsePredefinedJointAngles(ser.read_type_7("use_predefined_joint_angles"));
-      data.setPreset(ser.read_type_9("preset"));
-      ser.read_type_f("joint_angles", data.getJointAngles());
-      data.setTrajectoryDuration(ser.read_type_6("trajectory_duration"));
       data.setHoldPoseInWorld(ser.read_type_7("hold_pose_in_world"));
+      ser.read_type_a("screw_axis_pose", new controller_msgs.msg.dds.RigidBodyTransformMessagePubSubType(), data.getScrewAxisPose());
+
+      data.setTranslation(ser.read_type_6("translation"));
+      data.setRotation(ser.read_type_6("rotation"));
+      data.setMaxLinearVelocity(ser.read_type_6("max_linear_velocity"));
+      data.setMaxAngularVelocity(ser.read_type_6("max_angular_velocity"));
       data.setLinearPositionWeight(ser.read_type_6("linear_position_weight"));
       data.setAngularPositionWeight(ser.read_type_6("angular_position_weight"));
       data.setJointspaceWeight(ser.read_type_6("jointspace_weight"));
